remove userTypes from auth pages
