+++ src/App.js
@@ -4,8 +4,8 @@ import Header from './components/Header';
 import Footer from './components/Footer';
 import SearchPage from './pages/SearchPage';
 import AboutUs from './pages/AboutUs';
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
+import SignIn from './pages/authentication/SignIn';
+import SignUp from './pages/authentication/SignUp';
 import Splash from './pages/Splash';
 import Profile from './pages/Profile';
 import ExpandedCard from './pages/ExpandedCard';
