+++ src/pages/authentication/SignUp.jsx
@@ -4,54 +4,22 @@ import "./Sign.css";
 import "./desktop.css";
 
 export default function Auth() {
-  const [isHost, setIsHost] = useState(true);
   const [showPswd, setShowPswd] = useState(false);
-
-  const changeToHost = (e) => {
-    setIsHost(true);
-  };
-  const changeToGuest = () => {
-    setIsHost(false);
-  };
+  const [showModal, setShowModal] = useState(false);
 
   return (
     <div className="Container--centered fullHeight">
       <div className="Auth__container">
         <div className="Auth__heading">
           <i class="fas fa-arrow-left"></i>
-          <span>Sign In</span>
+          <span>Sign Up</span>
         </div>
 
         <div className="Auth__content">
-          <div className="Auth__userType">
-            <div
-              className={`${
-                isHost && "Auth__userTypeItem--selected"
-              } Auth__userTypeItem`}
-              onClick={changeToHost}
-            >
-              <span>
-                <i class="fas fa-user-shield" />
-              </span>
-              <span>I'm a host</span>
-            </div>
-            <div
-              className={`${
-                !isHost && "Auth__userTypeItem--selected"
-              } Auth__userTypeItem`}
-              onClick={changeToGuest}
-            >
-              <span>
-                <i class="fas fa-user" />
-              </span>
-              <span>I'm a guest</span>
-            </div>
-          </div>
-
           <UserForm
-            userType={isHost}
             showPswd={showPswd}
             setShowPswd={setShowPswd}
+            setShowModal={setShowModal}
           />
         </div>
       </div>
@@ -59,7 +27,7 @@ export default function Auth() {
   );
 }
 
-const UserForm = ({ userType, showPswd, setShowPswd }) => {
+const UserForm = ({ showPswd, setShowPswd }) => {
   const changeVisibility = () => {
     setShowPswd(!showPswd);
   };
@@ -67,10 +35,16 @@ const UserForm = ({ userType, showPswd, setShowPswd }) => {
     <>
       <form className="Auth__formItems">
         <div className="Auth__formItem">
-          <label for="name">{`${
-            userType ? "Host Email" : "Guest Email"
-          }`}</label>
+          <label for="name">Name</label>
           <input type="text" id="name" name="name" />
+          <label for="name">
+            Make sure it matches the name on your goverment ID.
+          </label>
+        </div>
+
+        <div className="Auth__formItem">
+          <label for="email">Email</label>
+          <input type="text" id="name" name="email" />
         </div>
 
         <div className="Auth__formItem">
@@ -84,13 +58,30 @@ const UserForm = ({ userType, showPswd, setShowPswd }) => {
           </div>
         </div>
 
-        <input type="submit" className="Auth__submitButton" value="Sign In" />
+        <div className="Auth__formItem">
+          <label for="name">Confirm Password</label>
+          <div name="password" className="Auth__formItem--password">
+            <input type={`${showPswd ? "text" : "password"}`} id="name" />
+            <i
+              className={`fas ${showPswd ? "fa-eye" : "fa-eye-slash"}`}
+              onClick={changeVisibility}
+            ></i>
+          </div>
+        </div>
+
+        <div className="Auth__TandC">
+          <input type="checkbox" name="tc" value="Bike" />
+          <label for="tc">
+            i agree to the Terms of service and Privacy policy
+          </label>
+        </div>
+
+        <input type="submit" className="Auth__submitButton" value="Sign Up" />
       </form>
 
       <div className="Auth--nav">
-        Don't have an account <Link to={"/Signup"}> Sign Up</Link>
+        Don't have an account <Link to={"/Signin"}> Sign In</Link>
       </div>
     </>
   );
 };
-
